+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+- DamagerIntersector component
 - DamageableFallback component
 
 ## [0.2.0] - 2025-11-17